Drop an unneeded mpz_get_ui() call
The loop above loops as long as the number is greater than zero so we
know that the result of the call will always be zero.  This is also
important because we were effectively processing a 0 as the most
significant digit.  It didn't cause us any problems because we handle
zeros in the loop directly and never look at hist[0].  However, it was
also a waste of valuable CPU cycles.